feat: :lipstick: Rearranged and add conditioning to show the Share and Copy btn
additional change: restructured the styling of sentiment and emotion analysis to be in a separate styling file
